Moves all debug stuff to the debug folder.
