+++ app/src/main/java/chat/rocket/android/RocketChatApplication.java
@@ -1,10 +1,7 @@
 package chat.rocket.android;
 
-import android.os.StrictMode;
 import android.support.multidex.MultiDexApplication;
 import com.crashlytics.android.Crashlytics;
-import com.facebook.stetho.Stetho;
-import com.uphyca.stetho_realm.RealmInspectorModulesProvider;
 
 import io.fabric.sdk.android.Fabric;
 import java.util.List;
@@ -19,13 +16,9 @@ import chat.rocket.persistence.realm.RocketChatPersistenceRealm;
  * Customized Application-class for Rocket.Chat
  */
 public class RocketChatApplication extends MultiDexApplication {
+
   @Override
   public void onCreate() {
-    if (BuildConfig.DEBUG) {
-      enableStrictMode();
-      enableStetho();
-    }
-
     super.onCreate();
     Fabric.with(this, new Crashlytics());
 
@@ -37,26 +30,5 @@ public class RocketChatApplication extends MultiDexApplication {
     }
 
     RocketChatWidgets.initialize(this, OkHttpHelper.getClientForDownloadFile(this));
-  }
-
-  private void enableStrictMode() {
-    StrictMode.setThreadPolicy(new StrictMode.ThreadPolicy.Builder()
-        .detectDiskReads()
-        .detectDiskWrites()
-        .detectNetwork()   // or .detectAll() for all detectable problems
-        .penaltyLog()
-        .build());
-    StrictMode.setVmPolicy(new StrictMode.VmPolicy.Builder()
-        .detectLeakedSqlLiteObjects()
-        .detectLeakedClosableObjects()
-        .penaltyLog()
-        .build());
-  }
-
-  private void enableStetho() {
-    Stetho.initialize(Stetho.newInitializerBuilder(this)
-        .enableDumpapp(Stetho.defaultDumperPluginsProvider(this))
-        .enableWebKitInspector(RealmInspectorModulesProvider.builder(this).build())
-        .build());
   }
 }
